rename entity package to be dto and CategoryEntity.kt to be CategoryDto.kt and TaskEntity.kt to be TaskDto.kt
