+++ app/src/main/java/com/amsterdam/cutetudee/data/mapper/CategoryMapper.kt
@@ -1,7 +1,7 @@
 package com.amsterdam.cutetudee.data.mapper
 
-import com.amsterdam.cutetudee.data.local.entity.CategoryEntity
-import com.amsterdam.cutetudee.data.local.entity.CategoryWithTaskCount
+import com.amsterdam.cutetudee.data.local.dto.CategoryDto
+import com.amsterdam.cutetudee.data.local.dto.CategoryWithTaskCount
 import com.amsterdam.cutetudee.domain.entity.Category
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
@@ -9,7 +9,7 @@ import kotlin.uuid.ExperimentalUuidApi
 import kotlin.uuid.Uuid
 
 @OptIn(ExperimentalUuidApi::class)
-fun Category.toCategoryEntity(): CategoryEntity = CategoryEntity(
+fun Category.toCategoryEntity(): CategoryDto = CategoryDto(
     id = id.toString(),
     name = name,
     imageUri = image,
