+++ app/src/test/java/com/amsterdam/cutetudee/data/service/CategoryServiceImplTest.kt
@@ -1,7 +1,7 @@
 package com.amsterdam.cutetudee.data.service
 
 import com.amsterdam.cutetudee.data.local.dao.CategoryDao
-import com.amsterdam.cutetudee.data.local.entity.CategoryWithTaskCount
+import com.amsterdam.cutetudee.data.local.dto.CategoryWithTaskCount
 import com.amsterdam.cutetudee.domain.model.Category
 import io.mockk.coEvery
 import io.mockk.coVerify
